ed8pkg2gltf.py: add support for relative source and destination path (#11)
* ed8pkg2gltf.py: add support for relative source and destination path

Before this commit all path were relative to the script location which
make sense when you execute the script from the windows explorer but
doesn't work well with relative path when executed from the command line
because it makes all path be relative to the script location not the
working directory of the interpreter.
Now the working directory is only changed to the script location if
there are no command line arguments.

* ed8pkg2gltf.py: remove unnecessary space