Add linux support to build_collada.py (#12)
* build_collada.py: make the generated shell script os independent

* revert unitentional changes

* fix incorrect path concatenation

* nit: fix metadata.json generation to work with -d flag

Before this change the pkg_name in the metadata.json would be the
full path to the pkg. Now it is only the pkg name.

* build_collada.py: fix shell script generation for animations

also contains a minor adjustment to the shell script header
on linux to make the script fail if any of its commands or
variable substitutions fail

---------

Co-authored-by: eArmada8 <103985972+eArmada8@users.noreply.github.com>