Set dynamic to 0 for physics rigid bodies
* ed8pkg2gltf.py: add support for relative source and destination path

Before this commit all path were relative to the script location which
make sense when you execute the script from the windows explorer but
doesn't work well with relative path when executed from the command line
because it makes all path be relative to the script location not the
working directory of the interpreter.
Now the working directory is only changed to the script location if
there are no command line arguments.

* ed8pkg2gltf.py: remove unnecessary space

* build_collada.py: always set dynamic to 0 for rigid bodies

Collision (in CS1) only works if dynamic is 0. I guess
this is because all rigit bodies in the .pkg are static
while dymaic rigid bodies are spawned via the scripts in
ops/ (this might be complete nonesense).

* Bugfix: remove j+1 line

---------

Co-authored-by: eArmada8 <103985972+eArmada8@users.noreply.github.com>