find_similar_shaders.py: add weights flags for kinds of differences
This change removes the -it and -nt flag in favor of a -w flag that
allows users to specify a dictionary of weights to give different types
of differences (e.g. texture_slot) different weights. This allows users
to specify they own metric for similarity and thus affect the ordering
in which the script returns the shaders (especially useful with the -nr
flag which return the most similar shader).

This also optimized the lookup to first filter the shaders based on the
game restrictions and then compute the similarity and order the shaders.
Before this change the script was computing and ordering all shaders and
then filtered afterwards.

I tested this on 50 random shaders for all games and the new and old
versions return exactly the same results for all of them (if -w is not
set).